# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: LibreRun
  version: 1.1.0-beta.1
paths:
...
  /api/v1/files/redact-preview:
    post:
      tags:
      - files
      summary: Redact Preview
      description: 'Redact an uploaded file and return the result for form prefill.


        Generalized in blueprint B8: any ``x-pii``-marked schema field can

        offer this preview — ``file_type`` selects the extension/size policy

        (``log`` or ``config``), and the old demo-agent-shaped ``vendor_side``

        parameter is gone (it never influenced the output).


        Answers ``503 pii_detector_unavailable`` when the detector is not

        ready (blueprint S4c): a preview whose whole promise is "this is

        what we would store" must not be able to show text that a broken

        stage 3 left whole.'
      operationId: redact_preview_api_v1_files_redact_preview_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_redact_preview_api_v1_files_redact_preview_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PiiRedactionPreview'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/runs/{run_id}/files:
    post:
      tags:
      - files
      summary: Upload Run File
      operationId: upload_run_file_api_v1_runs__run_id__files_post
      security:
      - HTTPBearer: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_run_file_api_v1_runs__run_id__files_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - files
      summary: List Run Files
      operationId: list_run_files_api_v1_runs__run_id__files_get
      security:
      - HTTPBearer: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/files/{file_id}:
    get:
      tags:
      - files
      summary: Download Run File
      operationId: download_run_file_api_v1_runs__run_id__files__file_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: File Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/runs/{run_id}/report/embedded:
    get:
      tags:
      - reports
      summary: Get Embedded Report
      description: 'Return the rendered report body as an HTML string for the React
        page.


        Tenant-scoped + owner-checked: only the case owner (or tenant admin)

        can fetch it. The returned HTML is the same template as the PDF/HTML

        download, with ``embedded=True`` (no ``<html>``/``<head>`` wrapper and

        a CSS prefix the agent''s template scopes its own styles under).'
      operationId: get_embedded_report_api_v1_runs__run_id__report_embedded_get
      security:
      - HTTPBearer: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/report:
    post:
      tags:
      - reports
      summary: Create Report
      operationId: create_report_api_v1_runs__run_id__report_post
      security:
      - HTTPBearer: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreate'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{task_id}:
    get:
      tags:
      - reports
      summary: Get Task
      operationId: get_task_api_v1_tasks__task_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{task_id}/download:
    get:
      tags:
      - reports
      summary: Download Task
      operationId: download_task_api_v1_tasks__task_id__download_get
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_redact_preview_api_v1_files_redact_preview_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        file_type:
          type: string
          title: File Type
          default: log
      type: object
      required:
      - file
      title: Body_redact_preview_api_v1_files_redact_preview_post
    Body_upload_run_file_api_v1_runs__run_id__files_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        vendor_side:
          type: string
          title: Vendor Side
        file_type:
          type: string
          title: File Type
      type: object
      required:
      - file
      - vendor_side
      - file_type
      title: Body_upload_run_file_api_v1_runs__run_id__files_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PiiRedactionPreview:
      properties:
        redacted_content:
          type: string
          title: Redacted Content
        redactions_applied:
          items:
            $ref: '#/components/schemas/RedactionEntry'
          type: array
          title: Redactions Applied
        original_size_bytes:
          type: integer
          title: Original Size Bytes
      type: object
      required:
      - redacted_content
      - redactions_applied
      - original_size_bytes
      title: PiiRedactionPreview
...
    RedactionEntry:
      properties:
        original_placeholder:
          type: string
          title: Original Placeholder
        pii_type:
          type: string
          title: Pii Type
        confidence:
          type: number
          title: Confidence
      type: object
      required:
      - original_placeholder
      - pii_type
      - confidence
      title: RedactionEntry
    ReportCreate:
      properties:
        format:
          type: string
          title: Format
          default: html
      type: object
      title: ReportCreate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer